# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScholarForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User deleted
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Projects
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects (public or member-accessible)
      parameters:
        - in: query
          name: visibility
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: myProjects
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted

  # Project Members
...
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /projects/{projectId}/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Task deleted

  # Milestones
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        institution:
          type: ["string", "null"]
        researchInterests:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        institution:
          type: ["string", "null"]
        researchInterests:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
      required:
        - role
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword

    # Project
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        abstract:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        visibility:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        memberCount:
          type: integer
        currentUserRole:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - keywords
        - status
        - visibility
        - createdAt
        - updatedAt
        - memberCount
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        visibility:
          type: string
        memberCount:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        currentUserRole:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - keywords
        - status
        - visibility
        - memberCount
        - taskCount
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        abstract:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        visibility:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
      required:
        - title
        - description
        - keywords
        - status
        - visibility
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        abstract:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        visibility:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]

    # Members
...
    Task:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        assignedToId:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        assignedTo:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required:
        - id
        - projectId
        - title
        - status
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        assignedToId:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
        - status
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        assignedToId:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]

    # Milestones
...